tgbot/pyproject.toml
[tool.poetry]
name = "tgbot"
version = "0.1.0"
description = ""
authors = ["Stanislav Semenov <semenov_sd@bk.ru>"]

[tool.poetry.dependencies]
python = "3.8.6"
aiogram = "^2.12"
SQLAlchemy = "^1.4.2"
loguru = "^0.5.3"
python-decouple = "^3.4"
APScheduler = "^3.7.0"
asyncpg = "^0.22.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
